Change the plugin to use LibRaw's dcraw_emu
LibRaw is more frequently updated than dcraw, DCRaw has not had a new
release since 2018.

The emulation that dcraw_emu provides is not complete, according to the
LibRaw documentation it is does not support for the -D, -d, -P, -K, -i
or -e command line switches.
It also adds a few new switches that support LibRaw features.